requirements.txt
google-generativeai
jupyterlab
python-dotenv
langchain-google-genai
mysql-connector-python